+++ ykdl/extractors/acfun/__init__.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+def get_extractor(url):
+    if '/bangumi/' in url:
+        from . import bangumi as s
+    else:
+        from . import video as s
+    return s.site
